feat: Watch specified plugin for changes
Implement before-watchPatterns hook that allows modification of the patterns that CLI uses to watch for file changes in the project.
When the hook returns a function, it automatically replaces the original method implementation. As other plugins, like nativescript-dev-webpack, nativescript-dev-typescript, nativescript-dev-sass, etc. use the same hook, the replacement receives the original method and can call it in case it decides to do it.
Similar to swizzling in Objective C, each of the plugin's hooks receives a "pointer" to the replaced method from the previous hook.
So for example in case nativescript-dev-typescript hook is called first, it will receive a "pointer" to the original method implementation and will replace it with its own function. This function may call the original method.
After that, our new hook will receive a "pointer" to nativescript-dev-typescript's function. When we call it , it will call the original method, modify the results and in our hook we'll receive the combination of both original and nativescrip-dev-typescript's patterns.
We get them, add our patterns and return the results.

In the current case, I've added the path to node_modules/nativescript-plugin to be watched by CLI. Also, I've removed the .ts files of the plugin from the watcher, so we'll not detect changes in them, CLI should be interested of transpiled files only.